% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blatent.R
\docType{package}
\name{blatent}
\alias{blatent}
\title{blatent: A package for estimating Bayesian latent variable models.}
\description{
Estimation of latent variable models using Bayesian methods. Currently supports diagnostic classification models.
}
