% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blatentPPMC.R
\name{blatentPPMC}
\alias{blatentPPMC}
\title{Simulates data using parameters from posterior distribution of blatent Markov chain}
\usage{
blatentPPMC(model, nSamples, seed = model$options$seed,
  type = c("covariance", "bivariate"))
}
\arguments{
\item{model}{A blatent MCMC model object.}

\item{nSamples}{The number of PPMC samples to be simulated.}

\item{seed}{The random number seed. Defaults to the seed set in the blatent model object.}

\item{type}{The type of statistic to generate. Defaults to c("covariance", "bivariate"), which produces covariances
and bivariate Chi-Square statistics. Not all of these may be good for model fit.}
}
\description{
Simulates data using parameters from posterior distribution of blatent Markov chain.
}
