% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDIC.R
\name{calculateDIC}
\alias{calculateDIC}
\title{calculateDIC}
\usage{
calculateDIC(model)
}
\arguments{
\item{model}{A blatent MCMC model object.}
}
\description{
Calculates DIC for a given model using model object specs.
}
