% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWAIC.R
\name{calculateWAIC}
\alias{calculateWAIC}
\title{calculateWAIC}
\usage{
calculateWAIC(model)
}
\arguments{
\item{model}{A blatent MCMC model object.}
}
\description{
Calculates WAIC for a given model using model object specs.
}
