% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDefaultPosteriorPredictiveChecks.R
\name{setDefaultPosteriorPredictiveChecks}
\alias{setDefaultPosteriorPredictiveChecks}
\title{Provides a list of posterior predictive model checks to be run following estimation of
       a Bayesian latent variable model.}
\usage{
setDefaultPosteriorPredictiveChecks(bivariate = TRUE,
  tetrachoric = TRUE)
}
\arguments{
\item{bivariate}{If \code{TRUE}, runs bivariate PPCMs producing the Pearson Chi-square statistic
for two-way tables comparing observed and simulated data for all pairs of items.
Defaults to \code{TRUE}.}

\item{tetrachoric}{If \code{TRUE}, calculates tetrachoric correlations between observed and simulated
data for all pairs of items. Defaults to \code{TRUE}.}
}
\value{
A list of named values containing a logical value for each parameter above.
}
\description{
Currently two types of posterior predictive model checks (PPMCs) are available: bivariate and
             tetrachoric.
}
